Now able to select a floor type!

+++ src/main/scala/org/oc/ld32/gui/GuiLabel.scala
@@ -1,6 +1,6 @@
 package org.oc.ld32.gui
 
-class GuiLabel(text: String, x: Int, y: Int, color: Int = 0xFFFFFFFF) extends BaguetteGui {
+class GuiLabel(text: String, x: Float, y: Float, color: Int = 0xFFFFFFFF) extends BaguetteGui {
   def render(delta: Float): Unit = {
     fontRenderer.renderString(text, x, y, color)
   }
